Move update-label actions, reducers to filter slice
